tokens/errors/lexer: small refactor and lift of `Location`
